User Modification, edit fixed

+++ src/Controller/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Post;
+use App\Entity\Image;
 use App\Form\PostType;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -12,33 +13,42 @@ use Symfony\Component\Routing\Annotation\Route;
 use DateTime;
 use App\Service\FileUploader;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
-use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
+use App\Security\PostVoter;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 
-#[Route('/posts')]
+#[Route('/post')]
 class PostsController extends AbstractController
 {
+    private $authorizationChecker;
+
+    public function __construct(AuthorizationCheckerInterface $authorizationChecker)
+    {
+        $this->authorizationChecker = $authorizationChecker;
+    }
 
     #[Route('/new', name: 'app_posts_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager, FileUploader $fileUploader, SessionInterface $session): Response
     {
+        // Vérifiez l'accès en utilisant le Voter
+        $this->denyAccessUnlessGranted(PostVoter::NEW , new Post());
+
         $post = new Post();
         $form = $this->createForm(PostType::class, $post);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $post->setUpdatedAt(new DateTime);
-            $post->setCreatedAt(new DateTime);
+
             $post->setAuthor($this->getUser());
-            // Accédez aux données du formulaire pour coverImgs et videos
-            $coverImgs = $form->get('coverImgs')->getData();
-            $array_imgs = [];
 
-            foreach ($coverImgs as $coverImg) {
-                $imageFileName = $fileUploader->upload($coverImg);
-                array_push($array_imgs, $imageFileName);
+            $images = $form->get('images')->getData();
+
+            foreach ($images as $image) {
+                $imageFilename = $fileUploader->upload($image->getUploadedFile());
+                $image->setLink($imageFilename);
+                $image->setPost($post);
+                $post->addImages($image);
             }
-
-            $post->setCoverImgs($array_imgs);
 
             $videos = $form->get('videos')->getData();
 
@@ -46,6 +56,11 @@ class PostsController extends AbstractController
                 $post->addVideos($video);
                 $video->setPost($post); // Assurez-vous que la vidéo est associée au post
             }
+
+            $mainImage = $form->get('mainImage')->getData();
+            $mainImageFileName = $fileUploader->upload($mainImage);
+            $post->setMainImage($mainImageFileName);
+
             $entityManager->persist($post);
             $entityManager->flush();
 
@@ -70,8 +85,8 @@ class PostsController extends AbstractController
     #[Route('/{id}/edit', name: 'app_posts_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Post $post, EntityManagerInterface $entityManager, FileUploader $fileUploader): Response
     {
-        // Récupérez les images existantes du post
-        $existingImagesData = $post->getCoverImgs();
+        // Vérifiez l'accès en utilisant le Voter
+        $this->denyAccessUnlessGranted(PostVoter::EDIT, $post);
 
         // Créez le formulaire en passant les données du post, y compris les images existantes
         $form = $this->createForm(PostType::class, $post);
@@ -80,52 +95,37 @@ class PostsController extends AbstractController
 
 
         if ($form->isSubmitted() && $form->isValid()) {
-            // Récupérez les nouvelles images du formulaire
-            $newCoverImgs = $form->get('coverImgs')->getData();
-
-            // Récupérez les index des images à supprimer depuis le champ caché du formulaire
-            $imagesToDeleteIndexes = $request->request->get('imagesToDeleteIndexes');
-
-            if (strlen($imagesToDeleteIndexes) > 0) {
-                $indexes = explode(',', $imagesToDeleteIndexes);
-
-                // Créez un tableau temporaire pour stocker les images à conserver
-                $tempExistingImagesData = [];
-
-                foreach ($existingImagesData as $index => $existingImage) {
-                    if (!in_array($index, $indexes)) {
-                        // L'index n'est pas dans le tableau des indices à supprimer, conservez l'image
-                        $tempExistingImagesData[] = $existingImage;
-                    } else {
-                        // Supprimez le fichier du système de fichiers en utilisant le service FileUploader
-                        $fileUploader->remove($existingImage);
-                    }
+            $mainImage = $form->get('mainImage')->getData();
+            if ($mainImage !== null) {
+                // Si une nouvelle image principale a été téléchargée, supprimez l'ancienne
+                $existingMainImage = $post->getMainImage();
+                if ($existingMainImage) {
+                    $fileUploader->remove($existingMainImage);
                 }
-
-                // Attribuez le tableau temporaire à $existingImagesData
-                $existingImagesData = $tempExistingImagesData;
+                // Téléchargez la nouvelle image principale et mettez à jour l'entité
+                $mainImageFileName = $fileUploader->upload($mainImage);
+                $post->setMainImage($mainImageFileName);
             }
 
-            // Traitez les nouvelles images comme suit
-            foreach ($newCoverImgs as $newCoverImg) {
-                // Téléchargez et ajoutez les nouvelles images au tableau
-                $imageFileName = $fileUploader->upload($newCoverImg);
-                $existingImagesData[] = $imageFileName;
+            $images = $form->get('images')->getData();
+            foreach ($images as $image) {
+                if ($image->getUploadedFile() !== null) {
+                    $imageFilename = $fileUploader->upload($image->getUploadedFile());
+                    $image->setLink($imageFilename);
+                    $image->setPost($post);
+                    $post->addImages($image);
+                }
             }
 
-            // Mettez à jour les images du post
-            $post->setCoverImgs($existingImagesData);
 
             $videos = $form->get('videos')->getData();
 
             foreach ($videos as $video) {
                 if ($video->getTitle() === null) {
-                    // La vidéo a un titre nul, nous la supprimons du tableau
                     $post->removeVideos($video);
                 } else {
-                    // La vidéo a un titre, nous l'ajoutons au post
                     $post->addVideos($video);
-                    $video->setPost($post); // Assurez-vous que la vidéo est associée au post
+                    $video->setPost($post);
                 }
             }
 
@@ -139,32 +139,46 @@ class PostsController extends AbstractController
             return $this->redirectToRoute('homepage', [], Response::HTTP_SEE_OTHER);
         }
 
+        // Récupérer tous les groupes uniques
+        $allPosts = $entityManager->getRepository(Post::class)->findAll();
+        $groups = ["Empty"];
+        foreach ($allPosts as $post) {
+            $group = $post->getGroupe();
+            if (!in_array($group, $groups)) {
+                $groups[] = $group;
+            }
+        }
+
         return $this->render('posts/edit.html.twig', [
             'post' => $post,
+            'allGroups' => $groups,
             'form' => $form->createView(),
         ]);
     }
 
 
-
-    
     #[Route('/{id}', name: 'app_posts_delete', methods: ['POST'])]
     public function delete(Request $request, Post $post, EntityManagerInterface $entityManager, FileUploader $fileUploader): Response
     {
+        // Vérifiez l'accès en utilisant le Voter
+        $this->denyAccessUnlessGranted(PostVoter::DELETE, $post);
+
         if ($this->isCsrfTokenValid('delete' . $post->getId(), $request->request->get('_token'))) {
             // Remove images associated with the post from the server
-            $coverImgs = $post->getCoverImgs();
-            foreach ($coverImgs as $coverImg) {
-                $fileUploader->remove($coverImg);
+            $images = $post->getImages();
+            foreach ($images as $image) {
+                $fileUploader->remove($image->getTitle());
             }
-    
+
             // Remove the post itself
             $entityManager->remove($post);
             $entityManager->flush();
         }
-    
+
         $this->addFlash('success', 'Suppression de la figure effectuée.');
         return $this->redirectToRoute('homepage', [], Response::HTTP_SEE_OTHER);
     }
-    
+
 }
+
+
